project5/requirements.txt
pandas
scipy
statsmodels
google-genai
requests
openpyxl
